Review of flatfield code, own compilation unit, use clip control value 0 when using auto clip control on already clipped image, code review
